requirements.txt
psycopg2==2.9.10
apache-airflow==2.5.1
requests==2.26.0
pandas==1.3.3
sqlalchemy==1.4.25
python-dotenv==0.19.1